This commit contains all kinds of stuff should be split into multiple commits. Some stuff needs to be cleaned up.
(See branch issues for other notes I made about this branch)

A small list of misc changes in this commit:
- Changed pieGetRescaling factor to float since all callers were dividing the result by 100.f
- Removed unimplemented sound function declarations.
- Disabled droid animations which are in the logical update
- Changed the lightmap pixel format to reduce upload time
- Change sphere projection method (this stuff needs cleaning up)
- Fixed mouse picking.

+++ src/effects.cpp
@@ -429,10 +429,10 @@ void initEffectsSystem(void)
 static void positionEffect(const EFFECT *psEffect)
 {
 	/* Establish world position */
-	Vector3i dv(
+	Vector3f dv(
 		psEffect->position.x - player.p.x,
 		psEffect->position.y,
-		psEffect->position.z - player.p.z
+		psEffect->position.z -  player.p.z
 	);
 
 	/* Push the indentity matrix */
